+++ ess-20220222/src/Models/ModifyEciScalingConfigurationRequest.php
@@ -251,6 +251,11 @@ class ModifyEciScalingConfigurationRequest extends Model
     public $ephemeralStorage;
 
     /**
+     * @var string
+     */
+    public $gpuDriverVersion;
+
+    /**
      * @description The hosts.
      *
      * @var hostAliases[]
@@ -502,6 +507,7 @@ class ModifyEciScalingConfigurationRequest extends Model
         'eipBandwidth'                  => 'EipBandwidth',
         'enableSls'                     => 'EnableSls',
         'ephemeralStorage'              => 'EphemeralStorage',
+        'gpuDriverVersion'              => 'GpuDriverVersion',
         'hostAliases'                   => 'HostAliases',
         'hostName'                      => 'HostName',
         'imageRegistryCredentials'      => 'ImageRegistryCredentials',
@@ -626,6 +632,9 @@ class ModifyEciScalingConfigurationRequest extends Model
         }
         if (null !== $this->ephemeralStorage) {
             $res['EphemeralStorage'] = $this->ephemeralStorage;
+        }
+        if (null !== $this->gpuDriverVersion) {
+            $res['GpuDriverVersion'] = $this->gpuDriverVersion;
         }
         if (null !== $this->hostAliases) {
             $res['HostAliases'] = [];
@@ -846,6 +855,9 @@ class ModifyEciScalingConfigurationRequest extends Model
         }
         if (isset($map['EphemeralStorage'])) {
             $model->ephemeralStorage = $map['EphemeralStorage'];
+        }
+        if (isset($map['GpuDriverVersion'])) {
+            $model->gpuDriverVersion = $map['GpuDriverVersion'];
         }
         if (isset($map['HostAliases'])) {
             if (!empty($map['HostAliases'])) {