+++ ess-20220222/ChangeLog.md
@@ -1,3 +1,12 @@
+2024-12-03 Version: 1.8.3
+- Update API CreateEciScalingConfiguration: add param GpuDriverVersion.
+- Update API CreateScalingGroup: add param CapacityOptions.
+- Update API DescribeEciScalingConfigurations: update response param.
+- Update API DescribeScalingGroups: update response param.
+- Update API ModifyEciScalingConfiguration: add param GpuDriverVersion.
+- Update API ModifyScalingGroup: add param CapacityOptions.
+
+
 2024-12-02 Version: 1.8.2
 - Update API DescribeElasticStrength: add param DataDiskCategories.
 - Update API DescribeElasticStrength: add param ImageFamily.